cleanup spacing etc.

+++ src/main/java/fi/metropolia/teemuerk/webstoreapi/mapper/OrderMapper.java
@@ -3,13 +3,11 @@ package fi.metropolia.teemuerk.webstoreapi.mapper;
 import fi.metropolia.teemuerk.webstoreapi.dto.OrderDto;
 import fi.metropolia.teemuerk.webstoreapi.dto.OrderItemDto;
 import fi.metropolia.teemuerk.webstoreapi.entity.*;
-
 import java.util.List;
 import java.util.stream.Collectors;
 
 public class OrderMapper {
 
-    // DTO → Entity
     public static Order toEntity(OrderDto dto, Customer customer, CustomerAddress address) {
         if (dto == null || customer == null) return null;
 
@@ -19,24 +17,20 @@ public class OrderMapper {
         order.setDelivery_date(dto.getDeliveryDate());
         order.setStatus(dto.getStatus());
         order.setShipping_address(address);
-
-
-
-
         return order;
     }
 
-    // Entity → DTO
+
     public static OrderDto toDto(Order order) {
         if (order == null) return null;
 
         OrderDto dto = new OrderDto();
-        dto.setId((Integer) order.getId());
-        dto.setCustomerId((Integer) order.getCustomer().getId());
+        dto.setId( order.getId());
+        dto.setCustomerId( order.getCustomer().getId());
         dto.setOrderDate(order.getOrder_date());
         dto.setDeliveryDate(order.getDelivery_date());
         dto.setStatus(order.getStatus());
-        dto.setShippingAddressId((Integer) order.getShipping_address().getId());
+        dto.setShippingAddressId( order.getShipping_address().getId());
 
         List<OrderItemDto> itemDtos = order.getItems().stream()
                 .map(OrderItemMapper::toDto)
